# A few optimizations