Improvements:
- Avoid allocating strings in progress bar
- Speed-up in computation of convergence error for specialized kSAT implementation


Benchmark:

## Job Properties
* Time of benchmarks:
    - Target: 8 Jan 2025 - 14:24
    - Baseline: 8 Jan 2025 - 14:26
* Package commits:
    - Target: 05a08f
    - Baseline: dbdf8b
* Julia commits:
    - Target: 5e9a32
    - Baseline: 5e9a32
* Julia command flags:
    - Target: None
    - Baseline: None
* Environment variables:
    - Target: None
    - Baseline: None

## Results
A ratio greater than `1.0` denotes a possible regression (marked with :x:), while a ratio less
than `1.0` denotes a possible improvement (marked with :white_check_mark:). All results are shown below.

| ID                             | time ratio                   | memory ratio                 |
|--------------------------------|------------------------------|------------------------------|
| `["generic bp", "run bp"]`     |                   0.98 (5%)  | 0.99 (1%) :white_check_mark: |
| `["generic bp", "run maxsum"]` |                   0.97 (5%)  |                   1.00 (1%)  |
| `["ising", "run bp"]`          | 0.94 (5%) :white_check_mark: | 0.83 (1%) :white_check_mark: |
| `["ising", "run maxsum"]`      | 0.87 (5%) :white_check_mark: | 0.90 (1%) :white_check_mark: |
| `["ksat", "generic"]`          |                   1.00 (5%)  |                   1.00 (1%)  |
| `["ksat", "optimized"]`        | 0.84 (5%) :white_check_mark: | 0.67 (1%) :white_check_mark: |

## Benchmark Group List
Here's a list of all the benchmark groups executed by this job:

- `["generic bp"]`
- `["ising"]`
- `["ksat"]`

## Julia versioninfo

### Target
```
Julia Version 1.11.2
Commit 5e9a32e7af2 (2024-12-01 20:02 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  uname: Linux 3.10.0-1160.15.2.el7.x86_64 #1 SMP Wed Feb 3 15:06:38 UTC 2021 x86_64 x86_64
  CPU: Intel(R) Xeon(R) CPU E5-2640 v4 @ 2.40GHz: 
                 speed         user         nice          sys         idle          irq
       #1-40  1575 MHz  2702748942 s  162750528 s  1164090554 s  4379887193 s          0 s
  Memory: 503.51097869873047 GB (312400.7578125 MB free)
  Uptime: 2.107763745e7 sec
  Load Avg:  6.8  7.82  8.88
  WORD_SIZE: 64
  LLVM: libLLVM-16.0.6 (ORCJIT, broadwell)
Threads: 1 default, 0 interactive, 1 GC (on 40 virtual cores)
```

### Baseline
```
Julia Version 1.11.2
Commit 5e9a32e7af2 (2024-12-01 20:02 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  uname: Linux 3.10.0-1160.15.2.el7.x86_64 #1 SMP Wed Feb 3 15:06:38 UTC 2021 x86_64 x86_64
  CPU: Intel(R) Xeon(R) CPU E5-2640 v4 @ 2.40GHz: 
                 speed         user         nice          sys         idle          irq
       #1-40  1201 MHz  2702757051 s  162750528 s  1164091901 s  4379928439 s          0 s
  Memory: 503.51097869873047 GB (314041.47265625 MB free)
  Uptime: 2.107776431e7 sec
  Load Avg:  7.26  7.57  8.65
  WORD_SIZE: 64
  LLVM: libLLVM-16.0.6 (ORCJIT, broadwell)
Threads: 1 default, 0 interactive, 1 GC (on 40 virtual cores)
```