@stecrotti 2025-01-08
# Benchmark Report for *BeliefPropagation*

## Job Properties
* Time of benchmarks:
    - Target: 8 Jan 2025 - 15:34
    - Baseline: 8 Jan 2025 - 15:36
* Package commits:
    - Target: ebd634
    - Baseline: dbdf8b
* Julia commits:
    - Target: 5e9a32
    - Baseline: 5e9a32
* Julia command flags:
    - Target: None
    - Baseline: None
* Environment variables:
    - Target: None
    - Baseline: None

## Results
A ratio greater than `1.0` denotes a possible regression (marked with :x:), while a ratio less
than `1.0` denotes a possible improvement (marked with :white_check_mark:). All results are shown below.

| ID                             | time ratio                   | memory ratio                 |
|--------------------------------|------------------------------|------------------------------|
| `["generic bp", "run bp"]`     |                   1.02 (5%)  | 0.99 (1%) :white_check_mark: |
| `["generic bp", "run maxsum"]` | 0.79 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["ising", "run bp"]`          | 0.95 (5%) :white_check_mark: | 0.83 (1%) :white_check_mark: |
| `["ising", "run maxsum"]`      |                   0.98 (5%)  | 0.90 (1%) :white_check_mark: |
| `["ksat", "generic"]`          | 0.91 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["ksat", "optimized"]`        | 0.79 (5%) :white_check_mark: | 0.67 (1%) :white_check_mark: |